+++ fiftyone/multimodal/server/routes.py
@@ -11,11 +11,12 @@ from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.responses import Response
 
-import fiftyone.multimodal.tags._temporal_tags as fota
+import fiftyone.core.tags as fota
 from fiftyone.multimodal.query import (
     resolve_playback_plan,
     resolve_scene_inventory,
 )
+from fiftyone.multimodal.schemas import v1 as foms
 from fiftyone.server import decorators
 from fiftyone.server.utils.datasets import get_dataset
 
@@ -220,7 +221,10 @@ def _temporal_tags_from_create_payload(
                 start=record.get("start", None),
                 end=record.get("end", None),
                 tag=record.get("tag", None),
-                index_type=record.get("index_type", fota.DEFAULT_INDEX_TYPE),
+                index_type=record.get(
+                    "index_type",
+                    foms.TimeTrackType.TIME_TRACK_TYPE_DURATION_NS,
+                ),
                 anchor=record.get("anchor", None),
                 kind=record.get("kind", None),
                 created_by=record.get("created_by", None),
